Make perldoc <url> use a reasonabe name
This patch makes perldoc with a URL use a somewhat reasonable name
based on the basename of the URL, instead of displaying PALLEGCSYO or
BJ1KKH1675 in the pod header (based on the tmp file name).

It treats all URLs the same (changing perldoc.pod to PERLDOC), instead
of treating .pm’s specially.  Maybe this could be improved later.

This is commit 8fe353ef0987121ae6b17c9dbd86962f54a703c0 in the perl repo
September 18, 2011