Add a pager that doesn't redirect stdin RT#85173
If perldoc is in the '-m' mode (show module source)
put the pager provided by the PERLDOC_SRC_PAGER at
the front of the list of pagers.

This is useful (for example) for loading a module
source in an editor without having to know the
full path to the module.