fix next head appearing in certain -f lookups
perlfunc.pod has groups of non-function keywords that direct users to the appropriate docs. Each of these are delimited by a `=head3`. If you look up one of the items in these lists, the next section's `=head3` shows up (for example, `AUTOLOAD`, `DESTROY`, `eq`).

This change stops gathering POD lines at the end of the list, instead of at the first item in the next list.