Move inspection of groff into main module (Perldoc.pm) from ToMan.pm:
The ToMan formatter would search for the groff version it has available
but based on CPAN RT #120229, we need to determine this much earlier to
decide whether to pass on to ToMan or not.

This does not make any behavioral changes (other than temporarily
removing a debug message).

* Two helper functions from Pod::Perldoc::BaseTo were moved into
  Perldoc.pm: _get_path_components() and _find_executable_in_path().
* Perldoc.pm now has a method inspect_execs() which tries to find all
  available executables of a given program. It uses data from a new
  helper function, _exec_data(). Currently only nroff is supported.
* From ToMan we removed the code for detecting nroff and instead
  incorporated them in _exec_data() or a new function
  _find_executable() -- both in Perldoc.pm.
* Perldoc.pm's _inspect_execs() now calls _find_executable(),
  _find_executable_in_path(), and _get_path_components().

ToTerm can now be switched to ToMan and everything still works as it did
before, but we are not doing this yet.

The next commit will start addressing the logic for picking ToMan
or ToTerm.