Clean up existing logic for determining whether to use ToTerm:
This is just too much to read or figure out at once. Instead of every
possible condition (including unless !(...), which is ridiculous), it's
simply the obvious cocnditions: A set of OS's or a check on term.

If I understand this logic correctly (and I had to write some tests to
verify), it means that we are not using ToTerm for Windows, DOS,
AmigaOS, or for the recognized terms listed ("*dumb*", "*emacs*",
"*none*", and "*unknown*". The fallback is ToText. I guess that makes
sense.