Find the best `less` from *all* available pagers:
We had assumed that the first `less` pager is of a sufficient version
but that is not necessarily the case.

Instead, we now go through all `less` pagers and test each one separately.
When we find one of the right version, we stop. If the user doesn't like
this, they should set their first pager to the right `less` pager.