Mandoc supports UTF-8 since 2011
The -Tlocale option has been the default since 2014 (and will remain the
default) while the -V option was removed in 2015 and won't come back.

All operating systems having a mandoc port or package ship a version
that is recent enough that UTF-8 just works, so testing for this is
unnecessary. Moreover, the actual test is broken since it would only
succeed on very outdated mandoc installations.