Add testing for pager guessing function:
Much of the future improvements require making sure we get the right
pager. To be able to do make the adjustments successfully, we need to
set a baseline with this test.

It tests environment variables, "-m" option, and OS options.

It already exposes a few options that might be worth creating tickets
for.