Guess pagers earlier:
We are trying to determine the formatter class (ToTerm, ToMan, etc.)
during `init()` but we only do the pager guessing (to figure out
which pagers we have) much later when we want to run the paging.

Moving the page guessing to the init point allows us to use the
pager availability and version to determine the formatter.

We also need to improve the detection of `less(1)` version and to
do that, we need to collect all possible pagers first, which is
done in `pagers_guessing()`.