Add test for the less version detected:
This change checks that the code for detecting the `less` version
without getting confused by arguments or redirections.

It tests every OS setup available, and checks whether it receives
the correct binary (after cleaning up redirection and arguments),
as well as the index of that binary, just so we know it's not the
same binary each time (if "less" is repeated, for example).

This regex has another character ("\s") that the original code
doesn't, to prove that it's correct to add it.

I also temporarily added a redirect to one of the "less" entries,
and it showed that the redirection regex works (though it broke
other tests, so I removed it before committing).