Verify we got less as a pager before adding -R to it:
As Zefram pointed out in a ticket, the `/less/` regexp match here doesn't
mean we didn't get a different pager with the string "less" in its name.

By checking that we received a version, we're also checking a regexp
match on the version call to return the result for "less" pager
(because the response to `--version` includes "less VERSION_NUMBER").

This means that now pagers that have the string "less" in their name
would not be tripped by it (unless they answer to `--version` with
"less VERSION_NUMBER") and `less` binaries that have the string "less"
but not only "less" would still work (like "lessng" if one would exist).

Arguably, we might want to try the `--version` check on all pagers
instead of ones that match `/less/` in their name, but... I'm not sure
we should.